Add a mapper to map system plugin response to SitePluginModel
